+++ lib/services/utils/logger/logger_util.dart
@@ -1,12 +1,17 @@
 import 'package:logger/logger.dart';
 
 class LoggerUtil {
-  static final Logger logger = Logger(
+  static Logger logger = Logger(
     level: Level.nothing,
     printer: PrettyPrinter(),
   );
 
   static void setLogLevel(Level level) {
-    Logger.level = level;
+    logger = Logger(
+      level: level,
+      printer: PrettyPrinter(
+        methodCount: 10,
+      ),
+    );
   }
 }